feat: clear all arc functions, update utils js
